Add files via upload
While the Iris data enabled a simple introduction to the k-means algorithm and cluster finding in general, most data sets will contain more features. As a result, we now switch to a real-world data with more features. Specifically, in this notebook we will apply the k-means algorithm to the credit card dataset which has 5 features. We will apply k-means on this dataset to develop a customer segmentation for marketing strategy